Restructure repo - need to fix imports
