+++ docs/README.md
@@ -1,0 +1,6 @@
+## /docs
+
+This directory contains the springtime documentation.
+
+`/src` contains the source code. Run `make html` inside this folder to generate
+the documentation locally. The output will be written to `/build`.
