+++ notebooks/README.md
@@ -1,0 +1,6 @@
+## /notebooks
+
+Use this folder for exploratory analysis or for the production of reports.
+Proposed naming convention:
+
+<initials>_<description>.ipynb (e.g. `pk_explore-google-earth-engine.ipynb`)
