+++ packages/README.md
@@ -1,0 +1,9 @@
+## /packages
+
+Use this directory for reusable (Python) code snippets. Python modules stored in
+these folders can be imported when working in the `../notebooks` and
+`../scripts` directory as if they were installed directly from PyPi.
+
+The `/springtime` directory is intended for project-specific code snippets. Once
+snippets start to turn into self-contained models, consider extracting them into
+their own top-level packages (such as `/py_ppo`).
