+++ packages/py_ppo/test_py_ppo.py
@@ -1,5 +1,5 @@
 import pytest
-from py_ppo import get_terms, download
+from packages.py_ppo.py_ppo import get_terms, download
 import numpy as np
 import pandas as pd
 import requests
