+++ packages/springtime/README.md
@@ -1,0 +1,5 @@
+## /packages/springtime
+
+Use this directory for project-specific code snippets. Once
+snippets start to turn into self-contained models, consider extracting them into
+their own top-level packages (such as `../py_ppo`).
