+++ scripts/README.md
@@ -1,0 +1,5 @@
+## /scripts
+
+Use this directory to store scripts for reproducing our results. Use
+[`click`](https://click.palletsprojects.com/en/latest/) so the scripts can be
+executed from the command line. If applicable, also add a rule to the makefile.
